requirements.txt
sentence-transformers==3.4.1
PyPDF2==3.0.1
mysql-connector-python==9.2.0
python-docx==1.1.2
textract==1.6.5
scikit-learn==1.6.1